Refactoring - Moved commonly used code to lib/common.
