+++ lib/common/ErrorCode.ts
@@ -1,0 +1,8 @@
+/**
+ * Error codes.
+ */
+enum ErrorCode {
+  InvalidTransactionNumberOrTimeHash = 'invalid_transaction_number_or_time_hash'
+}
+
+export default ErrorCode;
