+++ lib/common/ITransaction.ts
@@ -1,0 +1,9 @@
+/**
+ * Defines a Sidetree transaction.
+ */
+export default interface ITransaction {
+  transactionNumber: number;
+  transactionTime: number;
+  transactionTimeHash: string;
+  anchorFileHash: string;
+}
