+++ lib/core/IDidPublicKey.ts
