+++ src/core.ts
@@ -6,9 +6,9 @@ import * as getRawBody from 'raw-body';
 import * as Koa from 'koa';
 import * as Router from 'koa-router';
 import Core from '../lib/core/Core';
-import { IConfig } from '../lib/core/Config';
+import IConfig from '../lib/core/IConfig';
 import { IProtocolParameters } from '../lib/core/ProtocolParameters';
-import { IResponse, Response } from '../lib/core/Response';
+import { IResponse, Response } from '../lib/common/Response';
 
 /** Configuration used by this server. */
 interface IServerConfig extends IConfig {
